% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calheatmap.R
\name{calheatmapOutput}
\alias{calheatmapOutput}
\title{Widget output function for use in Shiny}
\usage{
calheatmapOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

