% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calheatmap.R
\name{renderCalheatmap}
\alias{renderCalheatmap}
\title{Widget render function for use in Shiny}
\usage{
renderCalheatmap(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

